src/python-worker/requirements.txt
librosa
numpy
Pillow
soundfile
tensorflow
asyncio
mariadb
aio-pika
noisereduce
audioread